refactor(indicateurs): regroupe le sujet drag/reorder dans un sous-dossier dedie
Deplace drag-handle, sortable-grid-row, sortable-group-body, use-grid-reorder
et son spec dans grid/drag-reorder/, comme keyboard-navigation/ et paste/.
Reecrit les chemins d'import ; aucun changement de logique.
