+++ apps/app/src/indicateurs/valeurs/grid/drag-reorder/drag-handle.tsx
